requirements.txt
# Core dependencies
google-adk>=0.5.0
google-generativeai>=0.8.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.30.0

# API and communication
httpx>=0.28.0
websockets>=13.0.0

# Data handling 
numpy>=2.0.0
pandas>=2.1.0

# Utils
typing-extensions>=4.0.0
tqdm>=4.65.0

# Security
authlib>=1.5.0
cryptography>=44.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
